TestCases

+++ ComarchCwiczenia/ComarchCwiczenia.Services/Invoices/InvoiceService.cs
@@ -9,7 +9,7 @@ namespace ComarchCwiczenia.Services.Invoices
 {
     public class InvoiceService
     {
-        public InvoiceItem CreateItem(string name, int netValue, int taxValue)
+        public InvoiceItem CreateItem(string name, decimal netValue, decimal taxValue)
         {
             InvoiceItem item = new()
             {
@@ -23,7 +23,7 @@ namespace ComarchCwiczenia.Services.Invoices
             return item;
         }
 
-        private decimal CalculateGross(int netValue, int taxValue)
+        private decimal CalculateGross(decimal netValue, decimal taxValue)
         {
             return netValue + (netValue * (taxValue / 100m));
         }
